refactor: split the tree row shell out of tree-rows
The deletion state and the jump digit pushed tree-rows.tsx to 521 lines,
past the cap. The seam it crossed was already there: one half answers what
ANY row looks like and how much width a label has left, the other answers
what a worktree / tab / routines / recent-jump row puts in that space.

tree-row-shell.tsx takes the first half — TreeRowShared, RowShell's marker
column and mouse contract, and the cell-budget helpers. 521 becomes 391 and
161, and every sidebar frame golden is byte-identical, so the move carries
no behavior with it.
