+++ packages/kobe/src/tui-react/panes/sidebar/SidebarTree.tsx
@@ -28,7 +28,7 @@ import { useLatest } from "../../lib/use-latest"
 import { ContextMenu } from "../../ui/context-menu"
 import { SidebarBrandHeader, SidebarCreateAction, SidebarNavRail, SidebarSearchInput, SidebarZenChip } from "./chrome"
 import { SidebarTreeBody } from "./tree-panel"
-import type { TreeRowShared } from "./tree-rows"
+import type { TreeRowShared } from "./tree-row-shell"
 import type { SidebarProps } from "./types"
 import { cursorTaskIdOf, useTreeBindings } from "./use-tree-bindings"
 import { useTreeMenu } from "./use-tree-menu"
